requirements.txt
web3==7.6.0
eth-typing>=3.5.0
eth-utils>=2.3.0
python-dotenv==1.0.0
pytest==7.4.3
pytest-asyncio==0.21.1


python-socketio[client]==5.11.1
websockets==12.0
aiohttp==3.9.3 
asyncio==3.4.3
